requirements.txt
PyMuPDF
scikit-learn
requests
python-dotenv
google-generativeai
streamlit